move viz methods to seperate module
