+++ src/trustyai/explainers/pdp.py
@@ -1,7 +1,5 @@
 """Explainers.pdp module"""
-
 import math
-import matplotlib.pyplot as plt
 import pandas as pd
 from pandas.io.formats.style import Styler
 
@@ -61,45 +59,6 @@ class PDPResults(ExplanationResults):
         Style object from the PDP pd.DataFrame (see as_dataframe)
         """
         return self.as_dataframe().style
-
-    def plot(self, output_name=None, block=True, call_show=True) -> None:
-        """
-        Parameters
-        ----------
-        output_name: str
-            name of the output to be plotted
-            Default to None
-        block: bool
-            whether the plotting operation
-            should be blocking or not
-        call_show: bool
-            (default= 'True') Whether plt.show() will be called by default at the end of
-            the plotting function. If `False`, the plot will be returned to the user for
-            further editing.
-        """
-        fig, axs = plt.subplots(len(self.pdp_graphs), constrained_layout=True)
-        p_idx = 0
-        for pdp_graph in self.pdp_graphs:
-            if output_name is not None and output_name != str(
-                pdp_graph.getOutput().getName()
-            ):
-                continue
-            fig.suptitle(str(pdp_graph.getOutput().getName()))
-            pdp_x = []
-            for i in range(len(pdp_graph.getX())):
-                pdp_x.append(self._to_plottable(pdp_graph.getX()[i]))
-            pdp_y = []
-            for i in range(len(pdp_graph.getY())):
-                pdp_y.append(self._to_plottable(pdp_graph.getY()[i]))
-            axs[p_idx].plot(pdp_x, pdp_y)
-            axs[p_idx].set_title(
-                str(pdp_graph.getFeature().getName()), loc="left", fontsize="small"
-            )
-            axs[p_idx].grid()
-            p_idx += 1
-        fig.supylabel("Partial Dependence Plot")
-        if call_show:
-            plt.show(block=block)
 
     @staticmethod
     def _to_plottable(datum: Value):
@@ -187,7 +146,7 @@ class _PredictionProviderMetadata:
         """
         return self.data.sample()
 
-    # pylint: disable = invalid-name
+    # pylint: disable = invalid-name, missing-final-newline
     @JOverride
     def getOutputShape(self):
         """
